Refactor to subcommand architecture
- Reorganize file structure for better modularity
- Convert greeting functionality to a subcommand
- Update tests to match new structure
- Prepare codebase for easier feature addition
